Change & test account/create api to POST requests
Add POST request handling to the account/create route. This route didn't
have a specific request type, and I changed it to only accept POST type
requests.

Additionally I added 2 tests which pass successfully when a POST request
is done with the required 'email' field and fail successfully when that
field is empty.