Add & test unique email validation
Add unique email validation to the existing POST request in
account/create. This is based on the Mongoose Unique Validation
package.

Successfully tested as can be seen when running 'npm run test'.

This commit creates a known issue which is when you run the test a
second time it will fail because there will be a duplicate email the
second time. The short term solution is to delete the document manually
with:
db.accounts.findOneAndDelete({"email": "alex11@test.com"})
when inside a mongo shell.
in the future this will be fixed with an after method in the test.