Add a notifications POST route and fix known issue
Fix of a known issue which was a part of the previous commit which is
that the test fails when inserting a duplicate entry to the accounts
collection. Mongoose documentation explained that I didn't need the
unique validation package and that mongoose validates uniqueness by
itself. Googling showed me that I just needed to switch to a mock
database and delete it before each test.

Add a new Schema file which creates a Notifications Schema. This schema
is used in a POST request which pushes into a Notifications collection.
Add a test which validates that this route works.